+++ src/contract/money/src/entrypoint/transfer_v1.rs
@@ -46,7 +46,7 @@ use crate::{
 /// `get_metadata` function for `Money::TransferV1`
 pub(crate) fn money_transfer_get_metadata_v1(
     _cid: ContractId,
-    call_idx: u32,
+    call_idx: u8,
     calls: Vec<DarkLeaf<ContractCall>>,
 ) -> Result<Vec<u8>, ContractError> {
     let self_ = &calls[call_idx as usize].data;
@@ -117,7 +117,7 @@ pub(crate) fn money_transfer_get_metadata_v1(
 /// `process_instruction` function for `Money::TransferV1`
 pub(crate) fn money_transfer_process_instruction_v1(
     cid: ContractId,
-    call_idx: u32,
+    call_idx: u8,
     calls: Vec<DarkLeaf<ContractCall>>,
 ) -> Result<Vec<u8>, ContractError> {
     let self_ = &calls[call_idx as usize];
